requirements.txt
pytest==9.0.2
pytest-xdist==3.8.0
python-dotenv==1.2.1
requests==2.32.5
selenium==4.40.0
webdriver-manager==4.0.2
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.1
supabase==2.3.4
httpx==0.24.1